requirements.txt
requests
python-dotenv
beautifulsoup4
python-slugify
openai
